Extract TransactionHelper
