+++ src/Tippy/Pages/Transactions/Details.cshtml
@@ -2,4 +2,4 @@
 @model Tippy.Pages.Transactions.DetailsModel
 @{ ViewData["Title"] = "Transaction"; }
 
-<h1 class="title">Transaction <span style="font-size:70%;">@Model.TxHash</span></h1>
+<h1 class="title">Transaction <span style="font-size:70%;">@Model.TransactionDetail.TransactionHash</span></h1>
