+++ src/Tippy/Pages/Transactions/Details.cshtml.cs
@@ -1,6 +1,11 @@
 using System;
+using System.Linq;
 using Ckb.Rpc;
+using Ckb.Types;
 using Microsoft.AspNetCore.Mvc;
+using Tippy.ApiData;
+using Tippy.Util;
+using static Tippy.ApiData.TransactionHelper;
 
 namespace Tippy.Pages.Transactions
 {
@@ -10,7 +15,7 @@ namespace Tippy.Pages.Transactions
         {
         }
 
-        public String TxHash = "";
+        public TransactionDetailResult TransactionDetail = default!;
 
         public IActionResult OnGet(string? txhash)
         {
@@ -24,11 +29,76 @@ namespace Tippy.Pages.Transactions
                 return NotFound();
             }
 
-            TxHash = txhash;
-
             Client client = new($"http://localhost:{ActiveProject.NodeRpcPort}");
 
-            // TODO
+            TransactionWithStatus? transactionWithStatus = client.GetTransaction(txhash);
+            if (transactionWithStatus == null)
+            {
+                return NotFound();
+            }
+
+            Transaction tx = transactionWithStatus.Transaction;
+
+            bool isCellbase = tx.Inputs[0].PreviousOutput.TxHash == TransactionHelper.EmptyHash;
+            string prefix = IsMainnet() ? "ckb" : "ckt";
+
+            TransactionDetailResult detail = new()
+            {
+                IsCellbase = isCellbase,
+                Witnesses = tx.Witnesses,
+                CellDeps = tx.CellDeps.Select(dep =>
+                {
+                    return new ApiData.CellDep()
+                    {
+                        DepType = dep.DepType,
+                        OutPoint = new ApiData.OutPoint()
+                        {
+                            TxHash = dep.OutPoint.TxHash,
+                            Index = Hex.HexToUInt32(dep.OutPoint.Index),
+                        }
+                    };
+                }).ToArray(),
+                HeaderDeps = tx.HeaderDeps,
+                TxStatus = transactionWithStatus.TxStatus.Status,
+                TransactionHash = txhash,
+                TransactionFee = "0", // "0" for cellbase
+                BlockNumber = "",
+                Version = "",
+                BlockTimestamp = "",
+            };
+
+            if (transactionWithStatus.TxStatus.BlockHash != null)
+            {
+                Block? block = client.GetBlock(transactionWithStatus.TxStatus.BlockHash);
+                if (block != null)
+                {
+                    Header header = block.Header;
+                    UInt64 blockNumber = Hex.HexToUInt64(header.Number);
+
+                    detail.BlockNumber = blockNumber.ToString();
+                    detail.Version = Hex.HexToUInt32(header.Version).ToString();
+                    detail.BlockTimestamp = Hex.HexToUInt64(header.Timestamp).ToString();
+
+                    var (displayInputs, displayOutputs) = GenerateCellbaseDisplayInfos(client, txhash, tx.Outputs, blockNumber, prefix);
+                    detail.DisplayInputs = displayInputs;
+                    detail.DisplayOutputs = displayOutputs;
+                }
+            }
+
+            if (!isCellbase)
+            {
+                var previousOutputs = GetPreviousOutputs(client, tx.Inputs);
+
+                UInt64 transactionFee = previousOutputs.Select(p => Hex.HexToUInt64(p.Capacity)).Aggregate((sum, cur) => sum + cur) -
+                    tx.Outputs.Select(o => Hex.HexToUInt64(o.Capacity)).Aggregate((sum, cur) => sum + cur);
+                detail.TransactionFee = transactionFee.ToString();
+
+                var (displayInputs, displayOutputs) = GenerateNotCellbaseDisplayInfos(tx.Inputs, tx.Outputs, previousOutputs, prefix, txhash);
+                detail.DisplayInputs = displayInputs;
+                detail.DisplayOutputs = displayOutputs;
+            }
+
+            TransactionDetail = detail;
 
             return Page();
         }
